HAAR-860: Added warning message, in case of user with no email id
