+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/service/EmailNotificationService.kt
@@ -4,7 +4,6 @@ import com.microsoft.applicationinsights.TelemetryClient
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.manageusersapi.resource.EmailNotificationDto
 import uk.gov.service.notify.NotificationClientApi
 import uk.gov.service.notify.NotificationClientException
