+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/service/UserService.kt
@@ -78,3 +78,18 @@ data class NomisUserDetails(
   @Schema(description = "Last name of the user", example = "Smith")
   val lastName: String,
 )
+
+data class EmailNotificationDto(
+  @Schema(description = "Username", example = "TEST_USER")
+  val username: String,
+
+  @Schema(description = "First name of the user", example = "John")
+  val firstName: String,
+
+  @Schema(description = "email of the user", example = "Smith@gov.uk")
+  val email: String?,
+
+  @Schema(description = "admin id who enabled user", example = "ADMIN_USR")
+  val admin: String
+
+)
